Change GameLoader implementation to take into account spaces and \n characters

+++ app/src/main/java/com/github/pedrovgs/androidgameboyemulator/core/gameloader/GameReader.java
@@ -24,7 +24,7 @@ public interface GameReader {
 
   void load(String gameUri) throws FileNotFoundException;
 
-  byte getByte() throws IOException;
+  int getByte() throws IOException;
 
   void closeGame() throws IOException;
 }
